Tweak mimicking, monster skill, and targeting systems
Mimic targeting now looks for an item within radius of the dummy if no characters are found, and chooses to shoot the dummy over allies if the attack priority is >= 1. Also fixed a bug in weapon mimicking, and tweaked the timer values to hopefully account for the different speeds of attack animations.